Fixing previous, can't derive on types, only structs, enums, etc.
Also added the derive on the ColumnType enum.

+++ src/common/mod.rs
@@ -4,9 +4,10 @@ use std::collections::HashMap;
 // Third party library imports
 use serde::{Deserialize, Serialize};
 
-pub const BLOCK_SIZE: u64 = 8192;
+pub const BLOCK_SIZE: usize = 8192;
 pub type Block = [u8; BLOCK_SIZE];
 
+#[derive(Serialize, Deserialize)]
 pub enum ColumnType {
     Int,   // i64
     Boolean,
@@ -15,14 +16,7 @@ pub enum ColumnType {
     Blob,   // for any size binary data
 }
 
-#[derive(Serialize, Deserialize)]
 pub type ColumnHeader = (String, ColumnType);
-
-#[derive(Serialize, Deserialize)]
 pub type TableInfo = Vec<ColumnHeader>;
-
-#[derive(Serialize, Deserialize)]
 pub type SchemaMap = HashMap<String, (ColumnType, u64)>;
-
-#[derive(Serialize, Deserialize)]
 pub type DBSchema = HashMap<String, TableInfo>;
